fix: fix previous PR notes
