+++ src/hooks/index.js
@@ -1,6 +1,6 @@
 import useAlert from './useAlert';
-import useDelay from './useDelay';
+import useModalToggling from './useModalToggling';
 import useLayoutEffect from './useLayoutEffect';
 import useDraggable from './useDraggable';
 
-export { useAlert, useDelay, useLayoutEffect, useDraggable };
+export { useAlert, useModalToggling, useLayoutEffect, useDraggable };
