+++ src/services/Authentication.js
@@ -1,7 +1,7 @@
 import firebase from 'firebase';
 import initializeService from '../utils/initializeService';
 import firebaseConfig from './firebase.config';
-import Azure from './Azure';
+import Heroku from './Heroku';
 
 const api = initializeService();
 const appForRegistration = firebase.initializeApp(firebaseConfig, 'Secondary');
@@ -24,7 +24,7 @@ export default class Authentication {
       if (isUserAddedToDb) {
         await this.secondaryAuthApp.createUserWithEmailAndPassword(email, password);
         await this.secondaryAuthApp.signOut();
-        const sendMailApi = api instanceof Azure ? api : new Azure();
+        const sendMailApi = api instanceof Heroku ? api : new Heroku();
         await sendMailApi.sendMailToUser(registrationData);
       }
     } catch ({ message }) {
